Completed BUGFIX
Issue was in ProductCard.jsx line 56:
Could not add 'dm' to className because product.id in className needed to stay unmutated for app.jsx line 95. Adding 'dm' changes the string and could not be used in relatedId variable.
Solution was to send the product.id through id attribute so 'dm' can be assigned to className.
In app.js, id was used for line 95 event.target.